fix: circular deps in store
